fix: not using HasSet.
Using HashSet here is a waste of memory. HashSet stands on HashMap, which will alloc amount of memory. The only usage is to travel the contents of the Set, ArraySet matches the scene.
Besides, it's a really bad practice to use inner class to "initialize" a set. It will produce more classes.